refactor: Move agent/command parsing to class methods
Move _parse_agent_md and _parse_command_md private utilities into
class methods on their respective BaseModel types:
- AgentDefinition.load(agent_path: Path)
- CommandDefinition.load(command_path: Path)

This makes the API more intuitive and follows the pattern of having
load methods on the data classes themselves.
